requirements.txt
# e-KreÅŸ Chatbot API â€” Dependencies
# ===================================

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data Validation
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Google Gemini AI
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9`r`nPyMuPDF>=1.27.0
